Finalise version range of sshd-sftp as [2,2.6.0)
Apache SSHD 2.6.0 changed the package of the class
"SftpSubsystemFactory". Fake SFTP Server Lambda uses this class and
therefore it is incompatible with sshd-sftp 2.6.0 and newer.
Setting version 2.6.0 as an exclusive upper bound of the version range
for sshd-sftp avoids that build tools choose an incompatible version and
break the tests of Fake SFTP Server Lambda's users.

There is also a new release 2.0.0 of Fake SFTP Server Lambda that is
compatible with Apache SSHD 2.6.0 and newer.