Fix compatibility issue with sshd-sftp 2.5.x
Version 2.5.0 of sshd-sftp added a new method "getUserHomeDir" to the
interface "FileSystemFactory". This change broke Fake SFTP Server Lambda
for users which use sshd-sftp 2.5.0 or newer because Fake SFTP Server
Lambda has an implementation of "FileSystemFactory" without the new
method.

This commit adds the method "getUserHomeDir" to the "DoNotCloseFactory"
(implementation of "FileSystemFactory"). Fake SFTP Server Lambda can now
be used with sshd-sftp 2.5.x and it can still be used with older 2.x
versions of sshd-sftp.