Add CloseableFakeSftpServer extends FakeSftpServer
The class can be instantiated, configured and closed manually, if doing
everything within a single lambda expression is undesirable for a
given use case. In order to achieve that, some methods and fields had to
be made protected (from final).

For example, in BDD (behaviour-driven development) users want to test in
a given-when-then fashion, i.e. separate the stimulus to the system
under test (when) from verifying results later (then). Some test
frameworks like Spock even require separate blocks, making the use of a
single lambda feel somewhat unnatural and clunky there.

Because the class implements AutoCloseable, it can be instantiated in a
try with resources structure.